SpotifyExtension
argosSpotifyExtension
This is a simple Spotify extension that displays the currently playing/paused song with artist, album and title. There are also some simple controls to pause/play, go to next or previous song and to exit spotify. If spotify isn't running, "Spotify is not running" is displayed together with a button to start Spotify.